requirements.txt
asgiref==3.6.0
git+https://github.com/getcohesive/marketplace-server-plugin-django.git@usage_tracking#egg=cohesive-marketplace-server-plugin-django
certifi==2022.12.7
charset-normalizer==3.0.1
contourpy==1.0.7
cycler==0.11.0
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
fonttools==4.38.0
gunicorn==20.1.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.3
numpy==1.24.1
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
whitenoise==6.3.0
djangorestframework-jwt
djangorestframework-simplejwt
